Added feature extraction and fit workflow unit test for base framwork. Added docstring when spotted. Added test_utils to store dummy data type.
